WIP try to separate how options are parsed to help locate the issue
